fragment firebase y recycler

+++ app/src/main/java/models/Pieza.kt
@@ -2,5 +2,5 @@ package models
 
 import android.widget.ImageView
 
-class Pieza(var Nombre: String?=null,var Marca: String?=null, var Modelo: String?=null, var Precio: Float?=null, var Descripcion: String? = null, var Foto: ImageView?=null) {
+class Pieza(var Nombre: String?=null,var Marca: String?=null, var Modelo: String?=null, var Precio: Double?=null, var Descripcion: String? = null, var Foto: ImageView?=null) {
 }
